+++ src/pages/organize/[orgId]/projects/[campId]/surveys/[surveyId]/insights.tsx
@@ -7,7 +7,7 @@ import EmptyOverview from 'features/surveys/components/EmptyOverview';
 import { getSurveyCampId } from 'features/surveys/utils/getSurveyUrl';
 import { PageWithLayout } from 'utils/types';
 import { scaffold } from 'utils/next';
-import ResponseStatsCards from 'features/surveys/components/ResponseStatsCards';
+import ResponseStatsCards from 'features/surveys/components/SurveyInsights';
 import SurveyLayout from 'features/surveys/layouts/SurveyLayout';
 import useServerSide from 'core/useServerSide';
 import useSurvey from 'features/surveys/hooks/useSurvey';
